requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
Faker==20.1.0

pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==7.4.3
pytest-django==4.7.0
